BE: Normalize `eth.s` asset
